Added Square and Triangle impl

+++ Bridge/src/figures/Figure.java
@@ -1,5 +1,5 @@
 package figures;
 
 public interface Figure {
-    int calculate();
+    double calculate();
 }
